Moved Pathfinding from Gameboard to Agent
